+++ examples/10-openapi/src/app/modules/routed/first/overriden-helper.ts
@@ -1,4 +1,4 @@
-import { Content, mediaTypeName, ContentOptions, Column } from '@ditsmod/openapi';
+import { Content, mediaTypeName, ContentOptions, Property } from '@ditsmod/openapi';
 
 /**
  * This class uses template to putting type model to `content`.
@@ -9,11 +9,11 @@ export class MetaContent extends Content {
   override set<T extends mediaTypeName = mediaTypeName>(contentOptions: ContentOptions<T>) {
     contentOptions = { ...contentOptions };
     class ApiResponse {
-      @Column({ type: 'array' }, contentOptions.model!)
+      @Property({ type: 'array' }, contentOptions.model!)
       data: any[];
-      @Column()
+      @Property()
       meta: any;
-      @Column()
+      @Property()
       error: any;
     }
     contentOptions.model = ApiResponse;