+++ examples/16-openapi-validation/src/app/modules/routed/first/models.ts
@@ -1,17 +1,17 @@
-import { Column, REQUIRED } from '@ditsmod/openapi';
+import { Property, REQUIRED } from '@ditsmod/openapi';
 
 export class Model1 {
-  @Column({ minimum: 3, maximum: 50, description: `ID should be between 3 and 50.` })
+  @Property({ minimum: 3, maximum: 50, description: `ID should be between 3 and 50.` })
   id: number;
 
-  @Column({ minLength: 3, maxLength: 10 })
+  @Property({ minLength: 3, maxLength: 10 })
   username: string;
 
-  @Column({ [REQUIRED]: true }, Number)
+  @Property({ [REQUIRED]: true }, Number)
   numbers: number[];
 }
 
 export class Model2 {
-  @Column()
+  @Property()
   model1: Model1;
 }